refactor: Rename convgru_modules.py to convgru.py
